+++ .gitignore
@@ -4,4 +4,4 @@
 /test
 /target
 /.vscode
-/todo
+/todo.md
