+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tron-lang"
-version = "2.3.0"
+version = "2.3.1"
 edition = "2021"
 authors = ["418e"]
 description = "Open Source Programming Language"
@@ -10,7 +10,4 @@ repository = "https://github.com/418e/Tron"
 license = "MIT"
 
 [dependencies]
-num = "0.4.1"
 rand = "0.8.5"
-colored = "2.0.4"
-hashbrown = "0.14.3"
