--- todo.md
@@ -1,22 +1,0 @@
-# Tron 3
-
-- complete CLI
-- complete STD
-- Production
-- Package Managment
-- Property Functions
-
-# Tron 4
-
-- optional static typing
-- interfaces and type declaration
-- enums, structs and matches
-- advanced math and libs
-
-# Tron 5
-
-- Compiler
-- Tron in Tron
-- Memory Managment
-- Async
-- Multithreading
